requirements.txt
nextcord==2.7.2
mysql-connector-python==8.0.34
python-dotenv==1.0.0
requests==2.31.0